Add files via upload
I often faced situations where I had a need to create multiple folders and initially had to do it manually until I found one of the powers of the .bat file. My most common need was to create a folder for each State of Nigeria and subfolders for the capital cities.  Here is how I did it: